+++ R/datasummary_correlation.R
@@ -130,6 +130,10 @@ datasummary_correlation <- function(data,
   sanitize_escape(escape) # after sanitize_output
   sanity_add_columns(add_columns)
   sanity_align(align)
+
+  if (inherits(data, "data.table")) {
+    data <- as.data.frame(data, check.names = FALSE)
+  }
   
   easycorrelation <- inherits(data, "easycorrelation")
 
@@ -239,6 +243,7 @@ datasummary_correlation <- function(data,
     add_columns = add_columns,
     notes = notes,
     title = title,
+    escape = escape,
     ...)
 
   # invisible return